refactor(Patches): Module renames for consistency
