+++ rust/src/patches/options.rs
