+++ rust/src/patches/strings.rs
