+++ rust/src/patches/vecs.rs
